Refactored Tabs component, created useCase file and added ReadMe
