--- src/packages/altTabs/altTabPanel.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export const AltTabPanel = () => {
-    return (
-        <div>
-            
-        </div>
-    )
-}
