--- src/packages/altTabs/useCaseAlt.js
@@ -1,16 +1,0 @@
-import { AltTab } from "./altTab";
-import { AltTabList } from "./altTabList";
-import { AltTabPanel } from "./altTabPanel";
-import { AltTabs } from "./altTabs";
-
-<AltTabs>
-  <AltTabList>
-    <AltTab>One</AltTab>
-    <AltTab>Two</AltTab>
-    <AltTab>Three</AltTab>
-  </AltTabList>
-
-  <AltTabPanel>
-    
-  </AltTabPanel>
-</AltTabs>
