+++ src/packages/tabs/src/Tabs.tsx
@@ -1,28 +1,28 @@
 import styled from "styled-components";
+import { MyStore } from "../context/myStore";
+import { useStore } from "../context/useStore";
 
 export const TabsComponent = styled.div`
   display: flex;
   flex-direction: column;
   gap: 30px;
   height: 50vh;
-
-  h1 {
-    border: 2px solid blue;
-  }
 `;
 
 /*  Tabs is wrapper component that holds list of tab topics and 
     contents for each tab in a different wrapper
-
-    Tab provides activeState variables to track what tab is open and display
-    its contents
 */
 
 interface Props {
-  children: any;
-  content: any;
+  children?: any;
 }
 
-export const Tabs: any = ({ children }: Props) => {
-  return <TabsComponent>{children}</TabsComponent>;
+export const Tabs = ({ children }: Props) => {
+  const myStore = useStore();
+  return (
+    // Store provider for hydrating tab content
+    <MyStore.Provider value={myStore}>
+      <TabsComponent>{children}</TabsComponent>;
+    </MyStore.Provider>
+  );
 };
